refacor: move route code to dedicated files
