+++ packages/server/src/index.ts
@@ -1,80 +1,27 @@
 import { Hono } from 'hono';
-import { getCookie, setCookie } from 'hono/cookie';
+import { getCookie } from 'hono/cookie';
 import { logger } from 'hono/logger';
-import * as exporter from 'exporter';
 import SwiggyError from 'exporter/SwiggyError';
 import { apiReference } from '@scalar/hono-api-reference';
 import openapiJson from '../public/openapi.json';
+import loginRoutes from './routes/login';
+import orderRoutes from './routes/swiggy-orders';
 
 const app = new Hono();
 app.use(logger());
 
 //OpenAPI
 app.get('/', apiReference({ spec: { content: openapiJson } }));
-app.get('/openapi', (c) => c.json(openapiJson));
 
-// Grab CSRF token and cookies for subsequent requests.
-app.get('/login', async (c) => {
-	const ua = c.req.header('User-Agent') ?? '';
+app.route('/login', loginRoutes);
+app.route('/orders', orderRoutes);
 
-	const data = await exporter.visitSwiggy(ua);
-	if (Object.keys(data).length) {
-		setCookie(c, 'request-cookies', data.requestCookies, { path: '/', httpOnly: true });
-		setCookie(c, 'request-csrf', data.csrf, { path: '/login', httpOnly: true });
-		return c.json({ status: 'Success', code: 0, message: 'Acquired CSRF token', data: data }, 200);
-	}
-	return c.json('Failed to get CSRF token from swiggy.com', 500);
-});
-
-// Generate otp
-app.post('/login/otp', async (c) => {
-	const ua = c.req.header('User-Agent') ?? '';
-	const body = await c.req.json();
-	const csrf = getCookie(c, 'request-csrf');
-	const requestCookies = getCookie(c, 'request-cookies');
-
-	if (!requestCookies || !csrf) {
-		throw new SwiggyError('csrf or requestCookies are invalid', 400);
-	} else {
-		await exporter.generateOTP(ua, body.mobileNumber, requestCookies, csrf);
-		return c.json(
-			{ status: 'Success', code: 0, message: 'Generated OTP Successfully', data: { csrf, requestCookies } },
-			200
-		);
-	}
-});
-
-// Login user
-app.post('/login/auth', async (c) => {
-	const ua = c.req.header('User-Agent') ?? '';
-	const body = await c.req.json();
-	const csrf = getCookie(c, 'request-csrf');
-	let requestCookies = getCookie(c, 'request-cookies');
-
-	if (!requestCookies || !csrf) {
-		throw new SwiggyError('csrf or requestCookies are invalid', 400);
-	} else {
-		requestCookies = await exporter.performLogin(ua, body.otp, requestCookies, csrf);
-		setCookie(c, 'request-cookies', requestCookies, { path: '/', httpOnly: true });
-		return c.json(
-			{ status: 'Success', code: 0, message: 'Logged in Sucessfully', data: { csrf, requestCookies } },
-			200
-		);
-	}
-});
-
-app.post('/orders', async (c) => {
-	const ua = c.req.header('User-Agent') ?? '';
-	const { lastOrderID, offSetID } = await c.req.json();
-
-	let requestCookies = getCookie(c, 'request-cookies');
-
-	if (!requestCookies) {
-		throw new SwiggyError('requestCookies are invalid', 400);
-	}
-	const data = await exporter.exportNewData(lastOrderID, requestCookies, ua, offSetID);
-
-	return c.json({ status: 'Success', code: 0, message: 'Fetched Orders Sucessfully', data: data }, 200);
+app.get('/health', (c) => {
+	return c.json({
+		status: 'ok',
+		uptime: process.uptime(),
+		timestamp: new Date().toISOString(),
+	});
 });
 
 app.onError((err, c) => {
@@ -107,4 +54,9 @@ app.onError((err, c) => {
 	);
 });
 
-export default app;
+export default {
+	port: Number(process.env.PORT) || 3000,
+	fetch: app.fetch,
+};
+
+console.log(`Server started on http://localhost:${process.env.PORT}`);
